+++ Wall.cpp
@@ -6,7 +6,7 @@ void Wall::Random()
 
 	_text = WALL_TEXT;
 }
-void Wall::CollidePoccess(int& point, Ball& ball)
+void Wall::CollidePoccess(float& point, Ball& ball)
 {
 	Obstacle::CollidePoccess(point, ball);
 }
